requirements.txt
# High-Performance HTTP/WebSocket Proxy - Dependencies
# Python 3.10+

fastapi>=0.109.0
starlette>=0.36.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
websockets>=12.0
python-dotenv>=1.0.0
python-json-logger>=2.0.0
